v0.3.10

+++ init.lua
@@ -11,8 +11,14 @@ local extensions = require("extensions")
 --[[ autoupdate ]]--
 do
 	local autoupdate = io.open("autoupdate", 'r') or io.open("autoupdate.txt", 'r')
-	if autoupdate then
-		autoupdate:close()
+	local semiupdate = not autoupdate and (io.open("semiautoupdate", 'r') or io.open("semiautoupdate.txt", 'r'))
+	if autoupdate or semiautoupdate then
+		if autoupdate then
+			autoupdate:close()
+		end
+		if semiupdate then
+			semiupdate:close()
+		end
 
 		coroutine.wrap(function()
 			local pkg = require("deps/fromage/package")
@@ -22,12 +28,24 @@ do
 				if lastVersion then
 					lastVersion = string.match(lastVersion, "version = \"(.-)\"")
 					if version ~= lastVersion then
-						for i = 1, #pkg.files do
-							os.remove("deps/fromage/" .. pkg.files[i])
+						local toUpdate
+						if semiupdate then
+							repeat
+								print("There is a new version of 'Fromage' available [" .. lastVersion .. "]. Update it now? (Y/N)")
+								toUpdate = string.lower(io.read())
+							until toUpdate == 'n' or toUpdate == 'y'
+						else
+							toUpdate = 'y'
 						end
-						os.execute("lit install Lautenschlager-id/fromage") -- Installs the new lib
-						os.execute("luvit " .. table.concat(args, ' ')) -- Luvit's command
-						return os.exit()
+					
+						if toUpdate == 'y' then
+							for i = 1, #pkg.files do
+								os.remove("deps/fromage/" .. pkg.files[i])
+							end
+							os.execute("lit install Lautenschlager-id/fromage") -- Installs the new lib
+							os.execute("luvit " .. table.concat(args, ' ')) -- Luvit's command
+							return os.exit()
+						end
 					end
 				end
 			end
